Update index.html
Refactor of the code to improve modularity, allowing me to work on it in a more piecemeal fashion.

Big update is the implementation of a world map. The AI can generate an 8x8 grid of locations, each with a name and description. The user can also edit this map. The user can select an adjacent location to move to, can set a destination far away to help track progress through the world, or can let the AI determine when you've moved to a new location. Locations can also have images associated with them, which allows the background image to be replaced as you enter the location.