Update index.html
Cleaned up the world map UI and fixed bugs.
Added AI gen buttons to more prompt fields.
Exposed each AI gen prompt in the settings modal in the prompt tab.
Added stop generation to the send button. Need to check that it fully cancels generation with koboldcpp.
Added location-specific memory. It is user-editable, but it will also be autogenerated when leaving a location.