Implemented AI Story Creation
## Huge update!
### AI Story Architect.
In the story library, there is a new story creation button at the top left. This begins the AI-powered story creation process that will craft a fully realized story, scenario, static knowledge, characters, and world map based on your prompt. Give it a try! (I can't promise this will work amazingly with local models, but failure at any step should not halt the process.)

### Gemini API *should* work now
Refined the Gemini API, adding what should be a dynamically pulled list of models from Google. Let me know if you experience issues with this.

### Options!
I added an option to disable the AI Analysis step that determines location changes and character sentiment. I also tried to improve the prompt to make it suck less, but we'll see.

I also added a toggle to show or hide the portrait in horizontal mode.

### Library Import
Gemini and I've fixed an issue with library imports that may have caused existing libraries to override imported libraries. We'll see if this fixes the issue.